Cleaned up API logic and separated all functions by concern.
