+++ backend/app/__init__.py
@@ -4,9 +4,8 @@ import os
 from app.database.db_init import init_db
 from app.routes.auth import auth_router
 from app.routes.db import db_router
-from app.routes.data import data_router
 from app.routes.llm import llm_router
-from app.routes.data_v2 import data_router2
+from app.routes.data import data_router
 
 load_dotenv()
 
@@ -15,9 +14,8 @@ app = FastAPI(title="AI Festival Recommender API")
 # Routes
 app.include_router(llm_router)
 app.include_router(auth_router)
-app.include_router(data_router)
 app.include_router(db_router)
-app.include_router(data_router2)
+app.include_router(data_router)
 
 
 @app.on_event("startup")
