+++ backend/app/routes/auth.py
@@ -8,8 +8,8 @@ import os
 
 auth_router = APIRouter(prefix="/auth", tags=["Auth"])
 
-# Initialize Spotify OAuth with configuration
 def get_spotify_oauth():
+    """Initializes Spotify OAUTH with configuration."""
     return SpotifyOAuth(
         client_id=os.getenv('CLIENT_ID'),
         client_secret=os.getenv('CLIENT_SECRET'),
@@ -17,15 +17,15 @@ def get_spotify_oauth():
         scope="user-library-read playlist-modify-private"
     )
 
-# Gain Spotify access token
 @auth_router.get("/login", response_class=RedirectResponse)
 async def login(sp_oauth: SpotifyOAuth = Depends(get_spotify_oauth)):
+    """Gathers Spotify access token to be used as login."""
     auth_url = sp_oauth.get_authorize_url()
     return RedirectResponse(url=auth_url)
 
-# Callback from Spotify login
 @auth_router.get("/callback")
 async def callback(code: str = Query(...), sp_oauth: SpotifyOAuth = Depends(get_spotify_oauth)):
+    """Callback functionality from Spotify login."""
     try:
         token_info = sp_oauth.get_access_token(code)
         return {"message": "Login successful!", "token_info": token_info}
