+++ backend/app/services/festival_service.py
@@ -1,89 +1,110 @@
+from app.models.stage.stage_artist import StageArtist
+from app.models.stage.stage_festival import StageFestival
+from app.models.stage.stage_tag import StageTag
 from sqlalchemy.ext.asyncio import AsyncSession
-from fastapi import Depends
-from sqlalchemy.future import select
-from sqlalchemy.sql import text
-from datetime import date, datetime
-from app.database.db_init import AsyncSessionLocal
-from app.models.festival import Festival
-from app.models.artist import Artist
-from app.models.tag import Tag
-from app.services.artist_service import save_artist_genre, create_or_get_artist, get_artists
-from app.services.tag_service import create_or_get_tag, get_tags
-from app.database.db_init import AsyncSessionLocal
-import re
+from datetime import datetime
+from app.services import utils, artist_service, llm_service
+import asyncio
 
-async def save_festival(festival_entry: dict):
-    async with AsyncSessionLocal() as db:
-        try:
-            print(f"✅ Attempting to save festival: {festival_entry['name']}")
-            festival, artists_added = await check_festival(festival_entry['name'], db)
-            
-            # Add new festival if does not exist
-            if festival is None:
-                print(f"✅ Creating new festival: {festival_entry['name']}")
-                festival = Festival(
-                    name=festival_entry['name'],
-                    location=festival_entry['location'],
-                    cancelled=festival_entry['cancelled'],
-                    start_date=festival_entry['start_date'],
-                    end_date=festival_entry['end_date']
-                )
+class TempFestival:
+    """Class to represent a festival object before inserting into the database."""
+    def __init__(self, name: str, location: str, date: datetime, link: str):
+        self.name = name
+        self.date = date
+        self.location = location
+        self.link = link
 
-                # for tag in festival_entry['tags']:
-                #     new_tag = await create_or_get_tag(tag, db)
-                #     # Create Link
-                #     festival.tags.append(new_tag)
-                all_tags = await get_tags(festival_entry['tags'], db)
-                
-
-            # # Check if artists updated
-            # if not artists_added and festival_entry['artists']:
-            #     for artist in festival_entry['artists']:
-            #         # accounts for b2b appearances (ex. Slander B2B Dimension)
-            #         for each_artist in artist.split(' B2B '):
-            #             new_artist, exists = await create_or_get_artist(each_artist, db)
-            #             if not exists:
-            #                 new_artists.append(new_artist.name)
-            #             # Create link, automatically inputted into intermediate table
-            #             if new_artist not in festival.artists:
-            #                 festival.artists.append(new_artist)
-                
-            #     await save_artist_genre(new_artists, db)
-
-            if (not artists_added) and festival_entry['artists']:
-                all_artists = await get_artists(festival_entry['artists'], db)
-                festival.artists.extend(all_artists)
-
-            print(f"✅ Attempring to add festival: {festival_entry['name']}")
-            db.add(festival)
-            await db.commit()
-            print(f"✅ Festival saved successfully: {festival_entry['name']}")
-        except Exception as e:
-            await db.rollback()
-            print(f"❌ Error during save: {e}")
+async def webscrape_all_festivals(db: AsyncSession):
+    """Scrapes all festival information from musicfestivalwizard."""
+    all_festivals = []
     
-# Function to remove all past festivals
-async def cleanup_past_festivals():
-    async with AsyncSessionLocal() as db:
-        try:
-            today = date.today()
-            result = await db.execute(select(Festival).filter(Festival.end_date < today))
-            past_festivals = result.scalars().all()
+    festival_url = "https://www.musicfestivalwizard.com/all-festivals/?festival_guide=us-festivals&festivalgenre=electronic"
+    soup = await utils.get_html(festival_url)
 
-            for festival in past_festivals:
-                festival.artists.clear()
-                festival.tags.clear()
-                print("deleting festival " + festival.name)
-                await db.delete(festival)
+    elements = soup.find_all("div", class_="entry-title search-title")
+    for element in elements:
+        # [name, location, date]
+        festival_info = element.get_text(separator='|', strip=True).split("|")
+        a_tag = element.find("a")
+        correct_date = lambda x: x.split('/')[0] if '/' in x else x
+        if a_tag:
+            page_link = a_tag["href"]
+            all_festivals.append(TempFestival(festival_info[0], festival_info[1], correct_date(festival_info[2]), page_link))
 
-            await db.commit()
+    # TODO: Check if festival already exists in database
 
-        except Exception as e:
-            await db.rollback()
-            print(f"An error occurred during cleanup: {e}")
+    tasks = [asyncio.create_task(scrape_festival_details(festival)) for festival in all_festivals]
+    await asyncio.gather(*tasks)
 
-# Helper function to parse start and end dates
-def parse_festival_date(date_str):
+    # TODO: deduplicate entries
+
+    for festival in all_festivals:
+        await save_temp_festival(festival, db)
+
+    await artist_service.batch_search_artists(50, db)
+
+    # join stage tables with main tables
+    await utils.merge_stage_tables(db)
+
+    # generate embeddings and put into table
+    await llm_service.generate_embeddings(all_festivals, db)
+
+    await db.commit()
+
+async def scrape_festival_details(festival: TempFestival):
+    """Scrapes detailed information on a festival."""
+    soup = await utils.get_html(festival.link)
+
+    # Finds all artist elements
+    artist_elements = soup.find_all("div", class_="lineupblock")
+    artists = [element.get_text(separator='|', strip=True).split("|") for element in artist_elements]
+    artists = artists[0] if artists else []
+    # TODO: find way to split b2b and vs artists
+    festival.artists = artists
+
+    # Finds all tag elements
+    tag_elements = soup.find_all("span", class_="heading-breadcrumb")
+    tags = [element.get_text(separator='|', strip=True).split("|") for element in tag_elements]
+    if tags:
+        tags = tags[0]
+    festival.tags = tags
+
+    # Processes proper dates
+    start_date, end_date = parse_festival_date(festival.date)
+    is_cancelled = start_date is None
+    if is_cancelled:
+        festival.artists = []
+    festival.start_date = start_date
+    festival.end_date = end_date
+    festival.cancelled = is_cancelled
+
+    
+
+async def save_temp_festival(festival: TempFestival, db: AsyncSession):
+    """Saves a temporary festival object to the database."""
+    new_festival = StageFestival(
+        name=festival.name, 
+        location=festival.location, 
+        cancelled=festival.cancelled,
+        start_date=festival.start_date, 
+        end_date=festival.end_date
+    )
+
+    for artist in festival.artists:
+        new_artist = await db.merge(StageArtist(name=artist))
+        if new_artist not in new_festival.stage_artists:
+            new_festival.stage_artists.append(new_artist)
+
+    for tag in festival.tags:
+        new_tag = await db.merge(StageTag(name=tag))
+        if new_tag not in new_festival.stage_tags:
+            new_festival.stage_tags.append(new_tag)
+
+    await db.merge(new_festival)
+    await db.commit()
+
+def parse_festival_date(date_str: str):
+    """Parses start and end dates from single string."""
     if date_str.strip().upper() == "CANCELLED":
         return None, None
     
@@ -110,18 +131,3 @@ def parse_festival_date(date_str):
     start_date = datetime.strptime(f"{start_month} {start_day} {year}", "%B %d %Y")
     end_date = datetime.strptime(f"{end_month} {end_day} {year}", "%B %d %Y")
     return start_date, end_date
-
-# Returns festival if exists AND if artists already added
-# returns festival (object), artists_added (boolean)
-# def check_festival(festival_name):
-#     festival = Festival.query.filter_by(name=festival_name).first()
-#     artists_added = bool(festival.artists) if festival is not None else False
-
-#     return festival, artists_added
-
-async def check_festival(festival_name: str, db: AsyncSession):
-    result = await db.execute(select(Festival).filter_by(name=festival_name))
-    festival = result.scalars().first()
-    artists_added = bool(festival.artists) if festival else False
-
-    return festival, artists_added
